Add initial instruction execution code
It does not work at this point, because `call_indirect` references function
that is not resolved. It turns out wasm-ld errors about undefined functions
are much more serious that I expected.

The problem is that I declared functions with `extern` and the appropriate
attribute, but wasm-ld still gave me `undefined reference`. So I enabled
undefined references with `-Wl,--allow-undefined`.

Now what I have is that when I try to run the module, it fails on
`call_indirect`.

It may turn out that using Emscripten is the only way to get this working
without reversing every bit of this mechanism.