Add line and column info to assembling errors and highlight lines in the source code
Z80Error has new members: `line` and `column`, which are returned to the
AssemblerWorker along with the exception string. AssemblerWorker converts
them to ExceptionInfo.

AssemblingError has a new `exceptions` member, which is an array of
ExceptionInfos.

Error lines are then retrieved in the main App component and sent to the
CodeEditor.

Remove error reporting from the MemoryView.